requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
asn1crypto==1.4.0
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
coverage==7.8.2
cryptography==42.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
durationpy==0.9
evdev==1.6.1
exceptiongroup==1.2.2
fastapi==0.109.2
fasteners==0.14.1
future==0.18.2
google-auth==2.27.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
iniconfig==2.1.0
iotop==0.6
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
kubernetes==29.0.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.2
olefile==0.46
packaging==25.0
paramiko==2.9.3
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.8.0
pillow==11.1.0
pluggy==1.6.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.7
pyparsing==2.4.7
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.0
pytest-cov==6.1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
systemd-python==234
tomli==2.2.1
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unified-logging==23.9.10.post488
urllib3==2.0.7
usb-creator==0.3.7
uvicorn==0.27.0.post1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.7.0
xdg==5
xkit==0.0.0
zipp==1.0.0
